backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
mysql-connector-python==8.3.0
requests==2.31.0
python-dateutil==2.8.2
